Review Page Clean-Up
Included changes:
* dispaly date of birth in a human-readable format;
* always display core service history period fields, even if they don't
have data; and
* responsive styling fixes for review page definition list.
