+++ src/applications/my-education-benefits/components/CustomReviewDOBField.jsx
@@ -1,11 +1,11 @@
 import React from 'react';
-import { formatReviewDate } from 'platform/forms-system/src/js/helpers';
+import { formatReadableDate } from '../helpers';
 
 export default function CustomReviewDOBField({ children }) {
   return (
     <div className="review-row">
       <dt>Your date of birth</dt>
-      <dd>{formatReviewDate(children?.props.formData)}</dd>
+      <dd>{formatReadableDate(children?.props.formData)}</dd>
     </div>
   );
 }
