+++ src/applications/my-education-benefits/helpers.js
@@ -66,6 +66,48 @@ export const unsureDescription = (
 );
 
 /**
+ * Formats a date in human-readable form. For example:
+ * January 1, 2000.
+ *
+ * @param {*} rawDate A date in the form '01-01-2000'
+ * @returns A human-readable date string.
+ */
+export const formatReadableDate = rawDate => {
+  const months = [
+    'January',
+    'February',
+    'March',
+    'April',
+    'May',
+    'June',
+    'July',
+    'August',
+    'September',
+    'October',
+    'November',
+    'December',
+  ];
+
+  let dateParts;
+  let date;
+
+  if (rawDate) {
+    dateParts = rawDate.split('-');
+    date = new Date(
+      Number.parseInt(dateParts[0], 10),
+      Number.parseInt(dateParts[1], 10) - 1,
+      Number.parseInt(dateParts[2], 10),
+    );
+  }
+
+  if (!date) {
+    return '';
+  }
+
+  return `${months[date.getMonth()]} ${date.getDate()}, ${date.getFullYear()}`;
+};
+
+/**
  * This function recieves the uiSchema and formData objects and returns
  * a comma separated list of the selected checkboxes in a checkbox
  * group.  The formData object contains the name of the checkbox
